adjust quota code location
