+++ src/app/backend/model/mysql/quota/quota.go
@@ -1,57 +1,11 @@
 package quota
 
 import (
-	mylog "app/backend/common/util/log"
 	mysql "app/backend/common/util/mysql"
 	localtime "app/backend/common/util/time"
 	"encoding/json"
 )
 
-var log = mylog.Log
-
-const (
-	QUOTA_SELECT = "SELECT id, name, cpu, mem, rbd, price, " +
-		"status, createdAt, modifiedAt, modifiedOp, comment " +
-		"FROM quota WHERE id=?"
-
-	QUOTA_SELECT_ALL = "SELECT id, name, cpu, mem, rbd, price, " +
-		"status, createdAt, modifiedAt, modifiedOp, comment " +
-		"FROM quota"
-
-	QUOTA_SELECT_ALL_ORDER_BY_CPU = "SELECT id, name, cpu, mem, rbd, price, " +
-		"status, createdAt, modifiedAt, modifiedOp, comment " +
-		"FROM quota " +
-		"ORDER BY cpu ASC"
-
-	QUOTA_INSERT = "INSERT INTO " +
-		"quota(name, cpu, mem, rbd, price, status, createdAt, modifiedAt, modifiedOp, comment) " +
-		"VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?)"
-
-	QUOTA_UPDATE = "UPDATE quota " +
-		"SET name=?, cpu=?, mem=?, rbd=?, price=?, status=?, modifiedAt=?, modifiedOp=?, comment=? " +
-		"WHERE id=?"
-
-	QUOTA_DELETE = "UPDATE quota " +
-		"SET status=?, modifiedAt=?, modifiedOp=? " +
-		"WHERE id=?"
-
-	VALID   = 1
-	INVALID = 0
-)
-
-type Quota struct {
-	Id         int32  `json:"id"`
-	Name       string `json:"name"`
-	Cpu        int32  `json:"cpu"`
-	Mem        int32  `json:"mem"`
-	Rbd        int32  `json:"rbd"`
-	Price      string `json:"price"`
-	Status     int32  `json:"status"`
-	CreatedAt  string `json:"createdAt"`
-	ModifiedAt string `json:"modifiedAt"`
-	ModifiedOp int32  `json:"modifiedOp"`
-	Comment    string `json:"comment"`
-}
 
 func NewQuota(name, price, comment string, cpu, mem, rbd, modifiedOp int32) *Quota {
 	return &Quota{
